+++ pg-ephemeral/src/config.rs
@@ -63,6 +63,7 @@ impl Instance {
             cross_container_access: self.cross_container_access,
             wait_available_timeout: self.wait_available_timeout,
             remove: true,
+            session_name: None,
             transparent_workdir: None,
         }
     }